Storyteller_Google_capstone/backend/requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
# google-adk is assumed to be installed or part of the local environment
